SPS cleanup
Former-commit-id: 6d01f05280487193748ab8cc4f0bf480b2793b58

+++ lib/headers/host_main_function.h
@@ -71,7 +71,7 @@ void main_function
 	float wide,
 	int	maxshift_original,
 	double	tsamp_original,
-	long int inc,
+	size_t inc,
 	float tstart,
 	float tstart_local,
 	float tsamp,
